Adds a filter for the device list view
An input field for an instant filter is displayed on top of the device list. Devices without a match in the name will be hidden while typing the filter...

For instance, by typing "Light", only devices matching "Light" in their name will be shown in the list.  The filter is case-insensitive.